feat: align backend with canonical API and schema

+++ server/src/main/java/com/ecm/server/dto/request/OrderItemValidateDto.java
@@ -22,5 +22,5 @@ public class OrderItemValidateDto {
 
     @NotNull(message = "Unit price is required")
     @Min(value = 0, message = "Unit price cannot be negative")
-    private Integer unitPrice;
+    private Long unitPrice;
 }
